move fixture to conftest so it can be reused
