+++ qcodes/tests/drivers/keysight_b1500/b1500_driver_tests/test_b1500.py
@@ -2,7 +2,6 @@ import warnings
 from unittest.mock import MagicMock
 
 import pytest
-from pyvisa import VisaIOError
 
 from qcodes.instrument_drivers.Keysight.keysightb1500 import constants
 from qcodes.instrument_drivers.Keysight.keysightb1500.constants import (
@@ -17,32 +16,6 @@ from qcodes.instrument_drivers.Keysight.keysightb1500.KeysightB1511B import B151
 from qcodes.instrument_drivers.Keysight.keysightb1500.KeysightB1517A import B1517A
 from qcodes.instrument_drivers.Keysight.keysightb1500.KeysightB1520A import B1520A
 from qcodes.instrument_drivers.Keysight.keysightb1500.KeysightB1530A import B1530A
-
-
-@pytest.fixture(name="b1500")
-def _make_b1500(request):
-    request.addfinalizer(KeysightB1500.close_all)
-
-    try:
-        resource_name = 'insert_Keysight_B2200_VISA_resource_name_here'
-        instance = KeysightB1500('SPA',
-                                 address=resource_name)
-    except (ValueError, VisaIOError):
-        # Either there is no VISA lib installed or there was no real
-        # instrument found at the specified address => use simulated instrument
-        import qcodes.instrument.sims as sims
-        path_to_yaml = sims.__file__.replace('__init__.py',
-                                             'keysight_b1500.yaml')
-
-        instance = KeysightB1500('SPA',
-                                 address='GPIB::1::INSTR',
-                                 visalib=path_to_yaml + '@sim'
-                                 )
-
-    instance.get_status()
-    instance.reset()
-
-    yield instance
 
 
 def test_make_module_from_model_name():
